+++ src/features/file-system/index.ts
@@ -1,0 +1,2 @@
+export * from "./components/FileSystemSelector";
+export * from "./utils/local-media";