+++ src/game/xstate/machines/mediaMachine.ts
@@ -1,45 +1,9 @@
 import { createMachine, assign, send } from "xstate";
-import * as Sentry from "@sentry/react";
 
-import { Severity } from "@/stores/notifications";
-import type { GameConfig } from "@/configureStore";
+import { createNotification, Severity } from "@/stores/notifications";
 import { MediaLink, MediaType } from "@/types/Media";
-import { createNotification } from "@/game/engine/notification";
-
-import fetchRedditPics from "../api/fetchRedditPics";
 
 const PRELOAD_LINK_THRESHOLD = 5;
-const ESTIMATED_SKIP_RATE = 1.5;
-const UPPER_LIMIT_CAP = 1000;
-const LOWER_LIMIT_CAP = 10;
-
-function getEstimatedRequiredLinkCount(
-  maximumGameTime: number,
-  slideDuration: number
-) {
-  let estimatedLinksRequired =
-    ((maximumGameTime * 60) / slideDuration) * ESTIMATED_SKIP_RATE;
-
-  // Cap the upper limit of the number of links to fetch.
-  estimatedLinksRequired = Math.min(estimatedLinksRequired, UPPER_LIMIT_CAP);
-
-  // Cap the lower limit of the number of links to fetch.
-  estimatedLinksRequired = Math.max(estimatedLinksRequired, LOWER_LIMIT_CAP);
-
-  // TODO: Unsure why, but sometimes the number is NaN.  Will investigate later once Sentry stops dropping events.
-  // For now if it's NaN let's set it to the lower limit cap to prevent an exception.
-  if (Number.isNaN(estimatedLinksRequired)) {
-    estimatedLinksRequired = LOWER_LIMIT_CAP;
-
-    Sentry.captureException(
-      new Error(
-        `getEstimatedRequiredLinkCount returns NaN, maximumGameTime: ${maximumGameTime}, slideDuration:${slideDuration}`
-      )
-    );
-  }
-
-  return Math.round(estimatedLinksRequired);
-}
 
 function preloadImage(url: string) {
   new Image().src = url;
@@ -66,12 +30,9 @@ export type MediaMachineEvent =
   | { type: "PRELOAD_LINK" }
   | StopEvent;
 
-export function createMediaMachine(config: GameConfig) {
-  const estimatedRequiredLinkCount = getEstimatedRequiredLinkCount(
-    config.gameDuration.max,
-    config.slideDuration
-  );
-
+export function createMediaMachine(
+  getLinks: () => Promise<MediaLink[] | undefined>
+) {
   const mediaMachine = createMachine<MediaMachineContext, MediaMachineEvent>(
     {
       id: "media",
@@ -92,12 +53,7 @@ export function createMediaMachine(config: GameConfig) {
         fetching: {
           invoke: {
             id: "fetchMedia",
-            src: () =>
-              fetchRedditPics({
-                subreddits: config.subreddits,
-                limit: estimatedRequiredLinkCount,
-                mediaTypes: config.imageType,
-              }),
+            src: getLinks,
             onDone: {
               target: "playing",
               actions: assign((context, event) => ({
@@ -106,7 +62,7 @@ export function createMediaMachine(config: GameConfig) {
             },
             onError: {
               target: "paused",
-              actions: (context, event) => {
+              actions: (_context, event) => {
                 createNotification({
                   message: `Error fetching media: ${event.data}.message}`,
                   duration: -1,